refactor: fixed entities package import after merge
